1.1.1. Correções de segurança e implementações menores. Foram feitas alterações para melhorar a usabilidae e deixar o sistema mais intuitivo, corrigindo falhas de segurança em acesso ao banco de dados.

+++ RestauranteOrlaJK/src/main/java/com/icaroelucas/restauranteorlajk/dto/NovoProdutoDTO.java
@@ -20,7 +20,7 @@ public class NovoProdutoDTO {
 		
 		Produto produto = new Produto();
 		produto.setNome(nome);
-		produto.setQuantidade(Integer.parseInt(quantidade));
+		produto.setQuantidade(Float.parseFloat(quantidade));
 		produto.setValidade(data);
 		produto.setMedida(Medida.valueOf(medida));
 		produto.setDisponivel(true);
